+++ stack.cpp
@@ -19,14 +19,14 @@ unsigned int pop(Stack* s){
     }
 }
 
-void pushEx(ExStack* s, float info) {
+void pushEx(ExStack* s, double info) {
     (s->topIndex)++;
     if(s->topIndex < MAX_NODES * MAX_ARITY){
         s->info[s->topIndex] = info;
     }
 }
 
-float popEx(ExStack* s) {
+double popEx(ExStack* s) {
     if(s->topIndex >= 0){
         (s->topIndex)--;
         return s->info[(s->topIndex) + 1];
